requirements.txt
streamlit
requests
pandas
folium
streamlit-folium
python-dotenv
workalendar
geopy
